Add missed files.
git-svn-id: https://trac.musala.com/ARDC/repo@10025 34708651-a5da-c34d-b20c-10ddab38317d

+++ src/com/musala/atmosphere/uiautomator/waitforexists/WaitForUIElementListener.java
@@ -30,7 +30,7 @@ public class WaitForUIElementListener implements Dispatchable {
     public void handle(Object[] args) throws UiObjectNotFoundException {
         UiElementDescriptor descriptor = (UiElementDescriptor) args[0];
 
-        Long timeout = (Long) args[1];
+        Integer timeout = (Integer) args[1];
 
         UiSelector selector = UiSelectorParser.convertSelector(descriptor);
         UiObject object = new UiObject(selector);
